Extract accounting capital account view model
