+++ src/Meridian.Ui/dashboard/src/screens/accounting-screen.view-model.test.ts
@@ -28,7 +28,6 @@ import {
   countOpenSecurityConflicts,
   resolveSecurityScheduleEvents,
   resolveSelectedReconciliation,
-  useCapitalAccountWorkbenchViewModel,
   useAccountingConfigurationViewModel,
   useAccountingReconciliationViewModel,
   useSecurityMasterViewModel
@@ -41,7 +40,6 @@ import {
 import type {
   AccountingConfigurationServices,
   AccountingReconciliationServices,
-  CapitalAccountWorkbenchServices,
   SecurityCashFlowScheduleEvent,
   SecurityMasterDrillInServices,
   SecurityMasterServices
@@ -53,10 +51,7 @@ import type {
   AccountingConfigurationWorkspace,
   AccountingProductionReadiness,
   AccountingTenantAdministrationProfile,
-  CapitalAccountWorkbench,
   RuleDryRunResult,
-  ManualJournalEntryDraft,
-  ManualJournalEntryWorkbench,
   LedgerJournalLine,
   LedgerTrialBalanceLine,
   ReconciliationCalibrationSummary,
@@ -789,740 +784,6 @@ const journalLines: LedgerJournalLine[] = [
     lineCount: 2
   }
 ];
-
-const manualJournalDraft: ManualJournalEntryDraft = {
-  journalEntryId: "manual-je-1",
-  status: "Draft",
-  fundProfileId: "fund-alpha",
-  ledgerBookId: "book-alpha",
-  accountingBasis: "Primary",
-  accountingDate: "2026-06-30",
-  periodId: "2026-06",
-  entityId: "entity-master",
-  fundNodeId: "fund-alpha",
-  currency: "USD",
-  memo: "Manual close adjustment",
-  preparedBy: "browser-user",
-  createdAtUtc: "2026-06-30T00:00:00Z",
-  updatedAtUtc: "2026-06-30T00:00:00Z",
-  version: 1,
-  lines: [
-    { lineId: "line-debit", side: "Debit", amount: 100, currency: "USD", accountPath: "Assets:Cash", securityId: null, securityDisplayName: null, description: "Cash debit" },
-    { lineId: "line-credit", side: "Credit", amount: 100, currency: "USD", accountPath: "Income:Interest", securityId: null, securityDisplayName: null, description: "Interest income credit" }
-  ],
-  evidenceLinks: [],
-  evidenceAttachments: [],
-  validationIssues: [
-    {
-      code: "manual-je.account-missing",
-      severity: "Critical",
-      message: "GL account was not found.",
-      targetId: "line-debit",
-      suggestedAction: "Choose an active GL account."
-    }
-  ],
-  totalDebits: 100,
-  totalCredits: 100,
-  imbalance: 0,
-  approvalId: null,
-  submittedAtUtc: null,
-  submittedBy: null,
-  entryType: "CapitalCall",
-  treasuryContext: {
-    effectiveDate: "2026-06-30",
-    idempotencyKey: "browser:fund-alpha:capital-call:manual-je-1",
-    fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-    fundEventType: "CapitalCall",
-    capitalAccountId: "capital-account:fund-alpha:lp-1",
-    investorId: "investor:lp-1",
-    paymentIntentId: "payment:fund-alpha:capital-call:manual-je-1",
-    settlementReference: "settlement:fund-alpha:capital-call:20260630"
-  }
-};
-
-const manualJournalWorkbench: ManualJournalEntryWorkbench = {
-  fundProfileId: "fund-alpha",
-  ledgerBookId: "book-alpha",
-  loadedAtUtc: "2026-06-30T00:00:00Z",
-  ledgerBooks: [],
-  chartOfAccounts: [
-    { nodeId: "cash", path: "Assets:Cash", accountName: "Cash", accountType: "Asset", isArchived: false },
-    { nodeId: "interest-income", path: "Income:Interest", accountName: "Interest Income", accountType: "Revenue", isArchived: false }
-  ],
-  drafts: [manualJournalDraft],
-  auditTrail: [],
-  privateCapitalActivity: {
-    fundProfileId: "fund-alpha",
-    ledgerBookId: "book-alpha",
-    projectedAtUtc: "2026-06-30T00:00:00Z",
-    fundEventCount: 1,
-    capitalAccountCount: 1,
-    submittedFundEventCount: 0,
-    approvalQueueCount: 0,
-    postedFundEventCount: 0,
-    publishedReportOutputCount: 0,
-    netCapitalActivity: 100,
-    currency: "USD",
-    fundEvents: [
-      {
-        fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-        fundEventType: "CapitalCall",
-        entryType: "CapitalCall",
-        journalStatus: "Draft",
-        journalEntryId: "manual-je-1",
-        effectiveDate: "2026-06-30",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        currency: "USD",
-        grossAmount: 100,
-        netCapitalActivity: 100,
-        memo: "Manual close adjustment",
-        paymentIntentId: "payment:fund-alpha:capital-call:manual-je-1",
-        settlementReference: "settlement:fund-alpha:capital-call:20260630",
-        evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-        validationIssues: [],
-        updatedAtUtc: "2026-06-30T00:00:00Z"
-      }
-    ],
-    capitalAccounts: [
-      {
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        currency: "USD",
-        contributions: 100,
-        distributions: 0,
-        subscriptions: 0,
-        redemptions: 0,
-        managementFees: 0,
-        netActivity: 100,
-        fundEventCount: 1,
-        lastEffectiveDate: "2026-06-30",
-        lastFundEventType: "CapitalCall",
-        fundEventIds: ["fund-event:fund-alpha:capital-call:20260630"]
-      }
-    ],
-    capitalAccountSubledgerEntries: [
-      {
-        subledgerEntryId: "capital-account-subledger:fund-event:fund-alpha:capital-call:20260630",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        currency: "USD",
-        fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-        fundEventType: "CapitalCall",
-        entryType: "CapitalCall",
-        approvalState: "Draft",
-        journalEntryId: "manual-je-1",
-        effectiveDate: "2026-06-30",
-        grossAmount: 100,
-        netCapitalActivity: 100,
-        runningNetActivity: 100,
-        memo: "Manual close adjustment",
-        evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-        validationIssues: [],
-        updatedAtUtc: "2026-06-30T00:00:00Z"
-      }
-    ],
-    ledgerImpacts: [
-      {
-        ledgerImpactId: "ledger-impact:fund-event:fund-alpha:capital-call:20260630:manual-je-1",
-        journalEntryId: "manual-je-1",
-        fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-        fundEventType: "CapitalCall",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        approvalState: "Draft",
-        effectiveDate: "2026-06-30",
-        currency: "USD",
-        totalDebits: 100,
-        totalCredits: 100,
-        imbalance: 0,
-        lineCount: 2,
-        isBalanced: true,
-        isPostingReady: false,
-        evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-        lines: [
-          {
-            lineId: "line-debit",
-            accountPath: "Assets:Cash",
-            side: "Debit",
-            amount: 100,
-            currency: "USD",
-            entityId: null,
-            securityId: null,
-            securityDisplayName: null,
-            evidenceLink: null
-          },
-          {
-            lineId: "line-credit",
-            accountPath: "Equity:Capital Contributions",
-            side: "Credit",
-            amount: 100,
-            currency: "USD",
-            entityId: null,
-            securityId: null,
-            securityDisplayName: null,
-            evidenceLink: null
-          }
-        ],
-        validationIssues: [
-          {
-            code: "manual-je.private-capital-ledger-impact-approval-pending",
-            severity: "Warning",
-            message: "Approval is pending.",
-            targetId: "manual-je-1",
-            suggestedAction: "Submit approval."
-          }
-        ]
-      }
-    ],
-    reportOutputs: [
-      {
-        reportOutputId: "report-output:fund-event:fund-alpha:capital-call:20260630:capitalcallnotice",
-        reportOutputType: "CapitalCallNotice",
-        displayName: "CapitalCallNotice for CapitalCall",
-        reportRoute: "/api/ledger/private-capital/activity?fundProfileId=fund-alpha&fundEventId=fund-event%3Afund-alpha%3Acapital-call%3A20260630&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1",
-        reportOutputRoute: "/api/ledger/private-capital/report-output?fundProfileId=fund-alpha&reportOutputId=report-output%3Afund-event%3Afund-alpha%3Acapital-call%3A20260630%3Acapitalcallnotice&fundEventId=fund-event%3Afund-alpha%3Acapital-call%3A20260630&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1",
-        fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-        fundEventType: "CapitalCall",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        approvalState: "Draft",
-        effectiveDate: "2026-06-30",
-        currency: "USD",
-        netCapitalActivity: 100,
-        evidenceLinkCount: 2,
-        evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-        isReportReady: false,
-        reportWorkflowState: "Draft",
-        reportLineProvenanceCount: 1,
-        readinessLabel: "Approval pending",
-        readinessReason: "Private-capital report output is not ready because the linked journal entry has not been submitted for approval.",
-        nextAction: "Submit or review approval",
-        nextActionRoute: "/api/ledger/journal-entry-workbench?fundProfileId=fund-alpha&journalEntryId=manual-je-1",
-        validationIssues: [
-          {
-            code: "manual-je.private-capital-report-approval-pending",
-            severity: "Warning",
-            message: "Approval is pending.",
-            targetId: "fund-event:fund-alpha:capital-call:20260630",
-            suggestedAction: "Submit approval."
-          }
-        ]
-      }
-    ],
-    fundEventRecords: [
-      {
-        fundEventRecordId: "fund-event-ledger-record:fund-event:fund-alpha:capital-call:20260630",
-        fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-        fundEventType: "CapitalCall",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        approvalState: "Draft",
-        journalEntryId: "manual-je-1",
-        effectiveDate: "2026-06-30",
-        currency: "USD",
-        grossAmount: 100,
-        netCapitalActivity: 100,
-        capitalAccountOpeningNetActivity: 0,
-        capitalAccountEndingNetActivity: 100,
-        memo: "Manual close adjustment",
-        paymentIntentId: "payment:fund-alpha:capital-call:manual-je-1",
-        settlementReference: "settlement:fund-alpha:capital-call:20260630",
-        activityRoute: "/api/ledger/private-capital/activity?fundProfileId=fund-alpha&fundEventId=fund-event%3Afund-alpha%3Acapital-call%3A20260630&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1",
-        evidenceRoute: "/api/workstation/evidence/subjects/private-capital-fund-event/fund-event%3Afund-alpha%3Acapital-call%3A20260630/packet",
-        approvalId: null,
-        approvalRoute: null,
-        isPosted: false,
-        isPostingReady: false,
-        isReportReady: false,
-        isPublished: false,
-        readiness: "ApprovalPending",
-        readinessLabel: "Approval pending",
-        readinessReason: "Submit the fund-event journal for approval before posting or stakeholder report output.",
-        nextAction: "Submit approval",
-        nextActionRoute: "/api/ledger/private-capital/activity?fundProfileId=fund-alpha&fundEventId=fund-event%3Afund-alpha%3Acapital-call%3A20260630&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1",
-        evidenceLinkCount: 2,
-        capitalAccountSubledgerEntryCount: 1,
-        ledgerImpactCount: 1,
-        reportOutputCount: 1,
-        validationIssueCount: 2,
-        primaryReportOutputId: "report-output:fund-event:fund-alpha:capital-call:20260630:capitalcallnotice",
-        primaryReportOutputType: "CapitalCallNotice",
-        primaryReportRoute: "/api/ledger/private-capital/report-output?fundProfileId=fund-alpha&reportOutputId=report-output%3Afund-event%3Afund-alpha%3Acapital-call%3A20260630%3Acapitalcallnotice&fundEventId=fund-event%3Afund-alpha%3Acapital-call%3A20260630&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1",
-        reportWorkflowState: "Draft",
-        publicationManifestId: null,
-        retainedManifestPath: null,
-        reportLineProvenanceCount: 1,
-        evidenceLinks: [
-          "/api/workstation/evidence/subjects/accounting-record/manual-je-1",
-          "/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"
-        ],
-        evidenceCategories: [
-          {
-            categoryId: "source-support",
-            label: "Source support",
-            isReady: true,
-            summary: "Source documents or retained evidence links support the fund event.",
-            evidenceLinkCount: 2,
-            evidenceLinks: [
-              "/api/workstation/evidence/subjects/accounting-record/manual-je-1",
-              "/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"
-            ],
-            requiredEvidence: ["Source document or retained evidence link"]
-          },
-          {
-            categoryId: "capital-account-subledger",
-            label: "Capital-account subledger",
-            isReady: true,
-            summary: "Capital-account impact is represented in the subledger.",
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            requiredEvidence: ["Capital-account impact"]
-          },
-          {
-            categoryId: "ledger-impact",
-            label: "Ledger impact",
-            isReady: false,
-            summary: "Balanced ledger impact and line evidence are available for the fund event.",
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            requiredEvidence: ["Balanced ledger impact", "Ledger line evidence"]
-          },
-          {
-            categoryId: "approval-state",
-            label: "Approval state",
-            isReady: false,
-            summary: "Approval reference is missing for the fund event.",
-            evidenceLinkCount: 0,
-            evidenceLinks: [],
-            requiredEvidence: ["Approval reference"]
-          },
-          {
-            categoryId: "payment-intent",
-            label: "Payment intent",
-            isReady: true,
-            summary: "Payment intent and settlement reference are captured.",
-            evidenceLinkCount: 2,
-            evidenceLinks: [
-              "payment:fund-alpha:capital-call:manual-je-1",
-              "settlement:fund-alpha:capital-call:20260630"
-            ],
-            requiredEvidence: ["Payment intent id", "Settlement reference"]
-          },
-          {
-            categoryId: "cash-evidence",
-            label: "Cash evidence",
-            isReady: true,
-            summary: "Payment intent payment:fund-alpha:capital-call:manual-je-1 and settlement settlement:fund-alpha:capital-call:20260630 have 1 retained cash evidence link(s); live execution remains deferred.",
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"],
-            requiredEvidence: []
-          },
-          {
-            categoryId: "report-output",
-            label: "Report output",
-            isReady: false,
-            summary: "Governed report output is linked to the fund event.",
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            requiredEvidence: ["Governed report output"]
-          }
-        ],
-        paymentIntentEvidence: {
-          paymentIntentId: "payment:fund-alpha:capital-call:manual-je-1",
-          settlementReference: "settlement:fund-alpha:capital-call:20260630",
-          status: "SettlementMatched",
-          isReady: true,
-          direction: "Inflow",
-          amount: 100,
-          currency: "USD",
-          effectiveDate: "2026-06-30",
-          summary: "Payment intent payment:fund-alpha:capital-call:manual-je-1 and settlement settlement:fund-alpha:capital-call:20260630 have 1 retained cash evidence link(s); live execution remains deferred.",
-          cashEvidenceLinkCount: 1,
-          cashEvidenceLinks: ["/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"],
-          requiredEvidence: [],
-          evidenceRoute: "/api/workstation/evidence/subjects/private-capital-fund-event/fund-event%3Afund-alpha%3Acapital-call%3A20260630/packet"
-        },
-        fundEvent: {
-          fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-          fundEventType: "CapitalCall",
-          entryType: "CapitalCall",
-          journalStatus: "Draft",
-          journalEntryId: "manual-je-1",
-          effectiveDate: "2026-06-30",
-          capitalAccountId: "capital-account:fund-alpha:lp-1",
-          investorId: "investor:lp-1",
-          currency: "USD",
-          grossAmount: 100,
-          netCapitalActivity: 100,
-          memo: "Manual close adjustment",
-          paymentIntentId: "payment:fund-alpha:capital-call:manual-je-1",
-          settlementReference: "settlement:fund-alpha:capital-call:20260630",
-          evidenceLinks: [
-            "/api/workstation/evidence/subjects/accounting-record/manual-je-1",
-            "/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"
-          ],
-          validationIssues: [],
-          updatedAtUtc: "2026-06-30T00:00:00Z"
-        },
-        capitalAccountSubledgerEntries: [
-          {
-            subledgerEntryId: "capital-account-subledger:fund-event:fund-alpha:capital-call:20260630",
-            capitalAccountId: "capital-account:fund-alpha:lp-1",
-            investorId: "investor:lp-1",
-            currency: "USD",
-            fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-            fundEventType: "CapitalCall",
-            entryType: "CapitalCall",
-            approvalState: "Draft",
-            journalEntryId: "manual-je-1",
-            effectiveDate: "2026-06-30",
-            grossAmount: 100,
-            netCapitalActivity: 100,
-            runningNetActivity: 100,
-            memo: "Manual close adjustment",
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            validationIssues: [],
-            updatedAtUtc: "2026-06-30T00:00:00Z"
-          }
-        ],
-        ledgerImpacts: [
-          {
-            ledgerImpactId: "ledger-impact:fund-event:fund-alpha:capital-call:20260630:manual-je-1",
-            journalEntryId: "manual-je-1",
-            fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-            fundEventType: "CapitalCall",
-            capitalAccountId: "capital-account:fund-alpha:lp-1",
-            investorId: "investor:lp-1",
-            approvalState: "Draft",
-            effectiveDate: "2026-06-30",
-            currency: "USD",
-            totalDebits: 100,
-            totalCredits: 100,
-            imbalance: 0,
-            lineCount: 2,
-            isBalanced: true,
-            isPostingReady: false,
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            lines: [
-              {
-                lineId: "line-debit",
-                accountPath: "Assets:Cash",
-                side: "Debit",
-                amount: 100,
-                currency: "USD",
-                entityId: null,
-                securityId: null,
-                securityDisplayName: null,
-                evidenceLink: null
-              },
-              {
-                lineId: "line-credit",
-                accountPath: "Equity:Capital Contributions",
-                side: "Credit",
-                amount: 100,
-                currency: "USD",
-                entityId: null,
-                securityId: null,
-                securityDisplayName: null,
-                evidenceLink: null
-              }
-            ],
-            validationIssues: [
-              {
-                code: "manual-je.private-capital-ledger-impact-approval-pending",
-                severity: "Warning",
-                message: "Approval is pending.",
-                targetId: "manual-je-1",
-                suggestedAction: "Submit approval."
-              }
-            ]
-          }
-        ],
-        reportOutputs: [
-          {
-            reportOutputId: "report-output:fund-event:fund-alpha:capital-call:20260630:capitalcallnotice",
-            reportOutputType: "CapitalCallNotice",
-            displayName: "CapitalCallNotice for CapitalCall",
-            reportRoute: "/api/ledger/private-capital/activity?fundProfileId=fund-alpha&fundEventId=fund-event%3Afund-alpha%3Acapital-call%3A20260630&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1",
-            fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-            fundEventType: "CapitalCall",
-            capitalAccountId: "capital-account:fund-alpha:lp-1",
-            investorId: "investor:lp-1",
-            approvalState: "Draft",
-            effectiveDate: "2026-06-30",
-            currency: "USD",
-            netCapitalActivity: 100,
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            isReportReady: false,
-            reportWorkflowState: "Draft",
-            reportLineProvenanceCount: 1,
-            readinessLabel: "Approval pending",
-            readinessReason: "Private-capital report output is not ready because the linked journal entry has not been submitted for approval.",
-            nextAction: "Submit or review approval",
-            nextActionRoute: "/api/ledger/journal-entry-workbench?fundProfileId=fund-alpha&journalEntryId=manual-je-1",
-            validationIssues: [
-              {
-                code: "manual-je.private-capital-report-approval-pending",
-                severity: "Warning",
-                message: "Approval is pending.",
-                targetId: "fund-event:fund-alpha:capital-call:20260630",
-                suggestedAction: "Submit approval."
-              }
-            ]
-          }
-        ],
-        validationIssues: [
-          {
-            code: "manual-je.private-capital-ledger-impact-approval-pending",
-            severity: "Warning",
-            message: "Approval is pending.",
-            targetId: "manual-je-1",
-            suggestedAction: "Submit approval."
-          },
-          {
-            code: "manual-je.private-capital-report-approval-pending",
-            severity: "Warning",
-            message: "Approval is pending.",
-            targetId: "fund-event:fund-alpha:capital-call:20260630",
-            suggestedAction: "Submit approval."
-          }
-        ]
-      }
-    ],
-    capitalAccountSubledgers: [
-      {
-        subledgerId: "capital-account-subledger:capital-account:fund-alpha:lp-1:investor:lp-1:usd",
-        fundProfileId: "fund-alpha",
-        ledgerBookId: "book-alpha",
-        projectedAtUtc: "2026-06-30T00:00:00Z",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        currency: "USD",
-        activityRoute: "/api/ledger/private-capital/capital-account-subledger?fundProfileId=fund-alpha&ledgerBookId=book-alpha&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1&currency=USD",
-        contributions: 100,
-        distributions: 0,
-        subscriptions: 0,
-        redemptions: 0,
-        managementFees: 0,
-        openingNetActivity: 0,
-        endingNetActivity: 100,
-        netCapitalActivity: 100,
-        fundEventCount: 1,
-        approvalQueueCount: 0,
-        postedFundEventCount: 0,
-        publishedReportOutputCount: 0,
-        evidenceLinkCount: 2,
-        validationIssueCount: 2,
-        firstEffectiveDate: "2026-06-30",
-        lastEffectiveDate: "2026-06-30",
-        lastFundEventType: "CapitalCall",
-        readiness: "ApprovalPending",
-        readinessLabel: "Approval pending",
-        readinessReason: "One or more fund events require approval before the capital-account subledger can be treated as posting ready.",
-        nextAction: "Submit or review approval",
-        nextActionRoute: "/api/ledger/private-capital/activity?fundProfileId=fund-alpha&fundEventId=fund-event%3Afund-alpha%3Acapital-call%3A20260630&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1",
-        evidenceLinks: [
-          "/api/workstation/evidence/subjects/accounting-record/manual-je-1",
-          "/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"
-        ],
-        evidenceCategories: [
-          {
-            categoryId: "source-support",
-            label: "Source support",
-            isReady: true,
-            summary: "Source support is retained for this capital account's fund events.",
-            evidenceLinkCount: 2,
-            evidenceLinks: [
-              "/api/workstation/evidence/subjects/accounting-record/manual-je-1",
-              "/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"
-            ],
-            requiredEvidence: ["Source document or retained evidence link"]
-          },
-          {
-            categoryId: "capital-account-subledger",
-            label: "Capital-account subledger",
-            isReady: true,
-            summary: "Capital-account impacts are represented in the running subledger.",
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            requiredEvidence: ["Capital-account impact"]
-          },
-          {
-            categoryId: "ledger-impact",
-            label: "Ledger impact",
-            isReady: false,
-            summary: "Ledger impacts are linked to this capital account.",
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            requiredEvidence: ["Balanced ledger impact", "Ledger line evidence"]
-          },
-          {
-            categoryId: "approval-state",
-            label: "Approval state",
-            isReady: false,
-            summary: "Approval references are missing for this capital account.",
-            evidenceLinkCount: 0,
-            evidenceLinks: [],
-            requiredEvidence: ["Approval reference"]
-          },
-          {
-            categoryId: "payment-intent",
-            label: "Payment intent",
-            isReady: true,
-            summary: "Payment intent and settlement reference are captured.",
-            evidenceLinkCount: 2,
-            evidenceLinks: [
-              "payment:fund-alpha:capital-call:manual-je-1",
-              "settlement:fund-alpha:capital-call:20260630"
-            ],
-            requiredEvidence: ["Payment intent id", "Settlement reference"]
-          },
-          {
-            categoryId: "cash-evidence",
-            label: "Cash evidence",
-            isReady: true,
-            summary: "1 payment intent(s), 1 settlement reference(s), and 1 retained cash evidence link(s) support this subledger; live execution remains deferred.",
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"],
-            requiredEvidence: []
-          },
-          {
-            categoryId: "report-output",
-            label: "Report output",
-            isReady: false,
-            summary: "Governed report outputs are linked to this capital account.",
-            evidenceLinkCount: 1,
-            evidenceLinks: ["/api/workstation/evidence/subjects/accounting-record/manual-je-1"],
-            requiredEvidence: ["Governed report output"]
-          }
-        ],
-        paymentIntentEvidence: {
-          paymentIntentId: "payment:fund-alpha:capital-call:manual-je-1",
-          settlementReference: "settlement:fund-alpha:capital-call:20260630",
-          status: "SettlementMatched",
-          isReady: true,
-          direction: "Inflow",
-          amount: 100,
-          currency: "USD",
-          effectiveDate: "2026-06-30",
-          summary: "1 payment intent(s), 1 settlement reference(s), and 1 retained cash evidence link(s) support this subledger; live execution remains deferred.",
-          cashEvidenceLinkCount: 1,
-          cashEvidenceLinks: ["/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"],
-          requiredEvidence: [],
-          evidenceRoute: "/api/ledger/private-capital/capital-account-subledger?fundProfileId=fund-alpha&ledgerBookId=book-alpha&capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1&currency=USD"
-        },
-        capitalAccount: null,
-        fundEventRecords: [],
-        subledgerEntries: [],
-        ledgerImpacts: [],
-        reportOutputs: [],
-        validationIssues: [
-          {
-            code: "manual-je.private-capital-ledger-impact-approval-pending",
-            severity: "Warning",
-            message: "Approval is pending.",
-            targetId: "manual-je-1",
-            suggestedAction: "Submit approval."
-          },
-          {
-            code: "manual-je.private-capital-report-approval-pending",
-            severity: "Warning",
-            message: "Approval is pending.",
-            targetId: "fund-event:fund-alpha:capital-call:20260630",
-            suggestedAction: "Submit approval."
-          }
-        ]
-      }
-    ],
-    paymentIntents: [
-      {
-        paymentIntentId: "payment:fund-alpha:capital-call:manual-je-1",
-        settlementReference: "settlement:fund-alpha:capital-call:20260630",
-        fundProfileId: "fund-alpha",
-        ledgerBookId: "book-alpha",
-        fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-        journalEntryId: "manual-je-1",
-        requester: "ops-user",
-        requestedAtUtc: "2026-06-30T00:00:00Z",
-        status: "ApprovalPending",
-        statusLabel: "Approval pending",
-        readinessReason: "Requester and expected movement are captured, but controller approval is not complete.",
-        executionDeferredReason: "Full payment execution is explicitly deferred in v0.18; this layer only retains intent, control, cash-evidence, reconciliation, and audit history before any bank-side instruction.",
-        expectedCashMovement: {
-          paymentIntentId: "payment:fund-alpha:capital-call:manual-je-1",
-          direction: "Inflow",
-          amount: 100,
-          currency: "USD",
-          effectiveDate: "2026-06-30",
-          settlementReference: "settlement:fund-alpha:capital-call:20260630",
-          fundEventId: "fund-event:fund-alpha:capital-call:20260630",
-          fundEventType: "CapitalCall",
-          capitalAccountId: "capital-account:fund-alpha:lp-1",
-          investorId: "investor:lp-1",
-          purpose: "Capital call for Fund Alpha LP",
-          payee: "fund:fund-alpha",
-          accountScope: "fund:fund-alpha / book:book-alpha / capital-account:fund-alpha:lp-1 / investor:lp-1",
-          businessPurpose: "Capital call for Fund Alpha LP",
-          approvalPolicy: "Controller approval pending before execution-deferred reliance",
-          sourceEvidenceLinks: [
-            "/api/workstation/evidence/subjects/accounting-record/manual-je",
-            "/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"
-          ]
-        },
-        evidenceRoute: "/api/workstation/evidence/subjects/payment-intent/payment%3Afund-alpha%3Acapital-call%3Amanual-je-1/packet",
-        workbenchRoute: "/api/ledger/private-capital/activity?fundProfileId=fund-alpha&paymentIntentId=payment%3Afund-alpha%3Acapital-call%3Amanual-je-1",
-        approvalChain: [
-          { sequence: 1, role: "Requester", actor: "ops-user", status: "Requested", decidedAtUtc: "2026-06-30T00:00:00Z", evidenceRoute: "/api/ledger/private-capital/activity?fundProfileId=fund-alpha" },
-          { sequence: 2, role: "Controller approval", actor: "controller", status: "Pending", decidedAtUtc: null, evidenceRoute: null }
-        ],
-        bankEvidence: [
-          {
-            evidenceId: "retained-cash-evidence:capital-call",
-            evidenceKind: "RetainedCashEvidence",
-            status: "Retained",
-            summary: "Retained wire evidence supports the expected cash movement.",
-            amount: 100,
-            currency: "USD",
-            effectiveDate: "2026-06-30",
-          recordedAtUtc: "2026-06-30T00:00:00Z",
-          externalRef: "settlement:fund-alpha:capital-call:20260630",
-          recordedBy: "cash-ops@example.com",
-          evidenceRoute: "/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"
-        }
-      ],
-        reconciliationLinks: [
-          {
-            linkId: "reconciliation:capital-call",
-            status: "Ready",
-            summary: "Cash evidence is linked to reconciliation review.",
-            evidenceRoute: "/api/reconciliation/runs/capital-call"
-          }
-        ],
-        auditHistory: [
-          {
-            auditEventId: "payment-intent-requested:manual-je-1",
-            recordedAtUtc: "2026-06-30T00:00:00Z",
-            actor: "ops-user",
-            action: "payment-intent.requested",
-            summary: "Payment intent was requested.",
-            evidenceLinks: ["/api/workstation/evidence/subjects/cash-evidence/payment-fund-alpha-capital-call-manual-je-1/packet"]
-          },
-          {
-            auditEventId: "payment-intent-execution-deferred:manual-je-1",
-            recordedAtUtc: "2026-06-30T00:00:00Z",
-            actor: "system",
-            action: "payment-intent.execution-deferred",
-            summary: "Payment execution remains deferred.",
-            evidenceLinks: []
-          }
-        ]
-      }
-    ],
-    validationIssues: []
-  }
-};
 
 describe("accounting-screen view model", () => {
   it("keeps Accounting task-mode routing outside the overloaded view model", () => {
@@ -3820,224 +3081,6 @@ describe("accounting-screen view model", () => {
     expect(result.current.createLedgerBookStatusText).toBe("Created Alpha Fund primary book.");
     expect(result.current.activateDisabledReason).toBeNull();
     expect(result.current.canActivate).toBe(true);
-  });
-
-  it("loads the Capital Account Workbench with investor evidence, allocation rules, lineage, and audit drill-through rows", async () => {
-    const workbench: CapitalAccountWorkbench = {
-      fundProfileId: "fund-alpha",
-      ledgerBookId: "book-alpha",
-      projectedAtUtc: "2026-06-30T17:00:00Z",
-      capitalAccountId: "capital-account:fund-alpha:lp-1",
-      investorId: "investor:lp-1",
-      currency: "USD",
-      workbenchRoute: "/api/ledger/private-capital/capital-account-workbench?capitalAccountId=capital-account%3Afund-alpha%3Alp-1",
-      statusLabel: "Restated lineage",
-      statusReason: "Statement lineage includes retained restatement metadata and audit evidence.",
-      investorAccountCount: 1,
-      fundEventCount: 1,
-      statementCount: 1,
-      restatementLineageCount: 1,
-      auditDrillThroughCount: 2,
-      netCapitalActivity: 125,
-      investorAccounts: [{
-        accountKey: "capital-account:fund-alpha:lp-1|investor:lp-1|USD",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        currency: "USD",
-        activityRoute: "/api/ledger/private-capital/capital-account-subledger?capitalAccountId=capital-account%3Afund-alpha%3Alp-1",
-        readiness: "Ready",
-        readinessLabel: "Ready",
-        readinessReason: "Retained evidence and statement support are available.",
-        nextAction: "Open statement",
-        nextActionRoute: "/api/ledger/private-capital/report-output?reportOutputId=report-output-1",
-        openingNetActivity: 0,
-        endingNetActivity: 125,
-        netCapitalActivity: 125,
-        contributions: 125,
-        distributions: 0,
-        subscriptions: 0,
-        redemptions: 0,
-        managementFees: 0,
-        fundEventCount: 1,
-        postedFundEventCount: 1,
-        approvalQueueCount: 0,
-        publishedReportOutputCount: 1,
-        evidenceLinkCount: 2,
-        validationIssueCount: 0,
-        evidenceCategorySummary: "2/2 allocation evidence categories ready.",
-        evidenceLinks: ["/evidence/source", "/evidence/cash"],
-        evidenceCategories: [],
-        fundEventRecords: [manualJournalWorkbench.privateCapitalActivity!.fundEventRecords[0]],
-        subledgerEntries: [],
-        ledgerImpacts: [],
-        reportOutputs: [],
-        validationIssues: [],
-        paymentIntentEvidence: {
-          paymentIntentId: "payment-1",
-          settlementReference: "settlement-1",
-          status: "SettlementMatched",
-          isReady: true,
-          direction: "Inflow",
-          amount: 125,
-          currency: "USD",
-          effectiveDate: "2026-06-30",
-          summary: "Cash evidence matched.",
-          cashEvidenceLinkCount: 1,
-          cashEvidenceLinks: ["/evidence/cash"],
-          requiredEvidence: [],
-          evidenceRoute: "/evidence/payment-1"
-        }
-      }],
-      allocationRules: [{
-        ruleId: "rule-source",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        categoryId: "source-support",
-        label: "Source support",
-        basis: "Fund event source support must be retained.",
-        isSatisfied: true,
-        reason: "Source support is retained.",
-        route: "/evidence/source",
-        evidenceLinkCount: 2,
-        evidenceLinks: ["/evidence/source"],
-        requiredEvidence: ["Source document"],
-        ruleVersion: "source-support:projection:1.1.1.1",
-        effectiveFrom: "2026-06-30",
-        effectiveTo: "2026-06-30",
-        formula: "fund_event.source_evidence_count > 0",
-        approvalState: "Approved",
-        approvalReference: "approval-lp-1 / /accounting/approvals?approvalId=approval-lp-1",
-        replayTrace: "Trace uses 1 fund event(s), 1 subledger entry(ies), 1 ledger impact(s), 1 report output(s), and 1 allocation input(s).",
-        inputs: [{
-          inputId: "allocation-input:fund-event:fund-event:fund-alpha:capital-call",
-          kind: "fund-event",
-          sourceId: "fund-event:fund-alpha:capital-call",
-          label: "CapitalCall / Capital call",
-          amount: 125,
-          currency: "USD",
-          effectiveDate: "2026-06-30",
-          evidenceRoute: "/evidence/source"
-        }],
-        relatedFundEventIds: ["fund-event:fund-alpha:capital-call"]
-      }],
-      statementLineage: [{
-        lineageId: "lineage-1",
-        capitalAccountId: "capital-account:fund-alpha:lp-1",
-        investorId: "investor:lp-1",
-        reportOutputId: "report-output-1",
-        reportOutputType: "CapitalAccountStatement",
-        displayName: "LP 1 Statement",
-        reportRoute: "/reporting/report-packs/lp-1",
-        reportPackId: "report-pack-1",
-        reportWorkflowState: "Restated",
-        isPublished: true,
-        isReportReady: true,
-        publicationManifestId: "manifest-1",
-        retainedManifestPath: "/evidence/manifest.json",
-        publicationEvidenceHash: "hash-1",
-        publishedAtUtc: "2026-06-30T17:00:00Z",
-        publishedBy: "publisher",
-        reportLineProvenanceCount: 3,
-        hasRestatementLineage: true,
-        restatementStatus: "Restatement lineage retained.",
-        restatementReasonCode: "capital-account-correction",
-        restatementPriorVersionReportId: "prior-report",
-        restatementApprover: "audit-partner",
-        restatementChangedLineCount: 1,
-        restatementEvidenceLinkCount: 1,
-        reportOutputRoute: "/api/ledger/private-capital/report-output?reportOutputId=report-output-1",
-        evidenceRoute: "/evidence/restatement",
-        capitalAccountSubledgerRoute: "/api/ledger/private-capital/capital-account-subledger?capitalAccountId=capital-account%3Afund-alpha%3Alp-1",
-        evidenceLinks: ["/evidence/report"],
-        restatementEvidenceLinks: ["/evidence/restatement"],
-        restatementChangedLines: [{
-          lineKey: "capital-account-ending",
-          previousValue: "100.00",
-          currentValue: "125.00",
-          evidenceLinkCount: 1,
-          evidenceLinks: ["/evidence/restatement"]
-        }]
-      }],
-      auditDrillThroughs: [{
-        drillThroughId: "drill-subledger",
-        kind: "subledger",
-        label: "LP 1 subledger",
-        summary: "Open capital-account subledger.",
-        route: "/api/ledger/private-capital/capital-account-subledger?capitalAccountId=capital-account%3Afund-alpha%3Alp-1",
-        isAvailable: true,
-        evidenceLinkCount: 2,
-        evidenceLinks: ["/evidence/source", "/evidence/cash"],
-        relatedIds: ["fund-event-1"]
-      }],
-      validationIssues: [],
-      liveCapabilities: ["Investor-level capital account evidence", "Statement publication and restatement lineage"],
-      plannedCapabilities: ["Full cap-table administration", "Broad LP portal self-service"]
-    };
-    const services: CapitalAccountWorkbenchServices = {
-      getWorkbench: vi.fn().mockResolvedValue(workbench)
-    };
-
-    const { result } = renderHook(() => useCapitalAccountWorkbenchViewModel(
-      true,
-      "?capitalAccountId=capital-account%3Afund-alpha%3Alp-1&investorId=investor%3Alp-1&currency=USD",
-      services
-    ));
-
-    await waitFor(() => expect(result.current.statusLabel).toBe("Restated lineage"));
-    expect(services.getWorkbench).toHaveBeenCalledWith(expect.objectContaining({
-      capitalAccountId: "capital-account:fund-alpha:lp-1",
-      investorId: "investor:lp-1",
-      currency: "USD"
-    }));
-    expect(result.current.summaryCards).toEqual(expect.arrayContaining([
-      expect.objectContaining({ id: "investor-accounts", value: "1" }),
-      expect.objectContaining({ id: "statements", detail: "1 restatement lineage" })
-    ]));
-    expect(result.current.investorAccounts[0]).toMatchObject({
-      title: "capital-account:fund-alpha:lp-1",
-      netActivityLabel: "+$125.00 USD",
-      paymentEvidenceLabel: "Settlement matched / Inflow / $125 USD / 1 cash evidence / settlement linked"
-    });
-    expect(result.current.fundEventCommandRows[0]).toMatchObject({
-      title: "CapitalCall",
-      readinessLabel: "Approval pending",
-      readinessReasonLabel: "Submit the fund-event journal for approval before posting or stakeholder report output.",
-      nextActionLabel: "Submit approval",
-      commandCenterRouteLabel: "/api/ledger/private-capital/fund-event-command-center?fundProfileId=fund-alpha&fundEventId=fund-event%3Afund-alpha%3Acapital-call%3A20260630",
-      evidenceRouteLabel: "/api/workstation/evidence/subjects/private-capital-fund-event/fund-event%3Afund-alpha%3Acapital-call%3A20260630/packet",
-      ledgerImpactLabel: "1 ledger impact(s)",
-      subledgerLabel: "1 subledger movement(s)",
-      reportOutputLabel: "1 report output(s)"
-    });
-    expect(result.current.allocationRules[0]).toMatchObject({
-      label: "Source support",
-      statusLabel: "Satisfied",
-      routeLabel: "/evidence/source",
-      policyLabel: "source-support:projection:1.1.1.1",
-      effectiveWindowLabel: "2026-06-30 -> 2026-06-30",
-      formulaLabel: "fund_event.source_evidence_count > 0",
-      approvalLabel: "Approved / approval-lp-1 / /accounting/approvals?approvalId=approval-lp-1",
-      inputSummaryLabel: "1 input(s): fund-event fund-event:fund-alpha:capital-call +$125.00 USD",
-      relatedFundEventLabel: "fund-event:fund-alpha:capital-call"
-    });
-    expect(result.current.statementLineage[0]).toMatchObject({
-      title: "LP 1 Statement",
-      statusLabel: "Restated",
-      restatementLabel: "capital-account-correction / 1 changed line(s) / 1 evidence",
-      changedLineRows: [
-        expect.objectContaining({
-          lineKey: "capital-account-ending",
-          valueLabel: "100.00 -> 125.00",
-          evidenceLabel: "1 changed-line evidence"
-        })
-      ]
-    });
-    expect(result.current.auditDrillThroughs[0]).toMatchObject({
-      title: "LP 1 subledger",
-      statusLabel: "Available"
-    });
-    expect(result.current.liveCapabilities).toContain("Investor-level capital account evidence");
-    expect(result.current.plannedCapabilities).toContain("Broad LP portal self-service");
   });
 
   it("derives canonical Accounting and Reporting loading states", () => {
